% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterConverter.R
\name{BetterConverter}
\alias{BetterConverter}
\title{Latitude and Longitude Converter}
\usage{
BetterConverter(input, type)
}
\arguments{
\item{input}{the vector of information to convert}

\item{type}{The type of data being entered ("lat" or "lon")}
}
\description{
This function takes latitude and longitude from minute format and converts to degrees; function requires parzer package
}
\keyword{latitude,}
\keyword{longitude}
